[Event Hubs] Fix browser bundling
