+++ sdk/eventhub/event-hubs/package.json
@@ -110,8 +110,8 @@
     "@azure/core-tracing": "^1.1.2",
     "@azure/core-util": "^1.9.1",
     "@azure/logger": "^1.0.0",
-    "buffer": "^6.0.0",
-    "is-buffer": "^2.0.3",
+    "buffer": "^6.0.3",
+    "is-buffer": "^2.0.5",
     "process": "^0.11.10",
     "rhea-promise": "^3.0.0",
     "tslib": "^2.6.3"
@@ -129,14 +129,13 @@
     "@rollup/plugin-inject": "^5.0.5",
     "@types/debug": "^4.1.4",
     "@types/node": "^18.0.0",
-    "@vitest/browser": "^2.0.4",
-    "@vitest/coverage-istanbul": "^2.0.4",
+    "@vitest/browser": "^2.0.5",
+    "@vitest/coverage-istanbul": "^2.0.5",
     "@types/chai-as-promised": "7.1.8",
-    "@types/chai-string": "^1.4.5",
     "@types/ws": "^7.2.4",
+    "chai": "^5.1.1",
     "chai-as-promised": "^8.0.0",
-    "chai-exclude": "^2.1.1",
-    "chai-string": "^1.5.0",
+    "chai-exclude": "^3.0.0",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
     "debug": "^4.1.1",
@@ -149,7 +148,7 @@
     "tshy": "^3.0.2",
     "tsx": "^4.16.2",
     "typescript": "~5.5.4",
-    "vitest": "^2.0.4",
+    "vitest": "^2.0.5",
     "ws": "^8.2.0"
   },
   "exports": {
